new: add build infrastructure, first target: lint
add a build infrastructure, makes commands from each .py file and does
something with them, using 'make' as usual.

first user of build system is the linting process.  we add rcfiles for
both flake8 (pyflakes + pep8) and pylint3, which gives us quite good
coverage.  rcfiles disable some of the checks we aren't willing to
change because they're silly and/or we'll do it later when needed (such
as adding docstrings)