/ installx: main: don't trap bubbled exceptions within subprogram
if we run the subprogram itself under 'try' then any uncaught exception in
the entire thing will bubble up and we'll catch it, which we don't want.
we actually want the exception catch localized because we're just
looking up the function name in the global namespace and want the
KeyError and TypeError to limit themselves to that application only.
otherwise we get "no such invocation" when these errors happen anywhere
in installx()/installrc() and aren't directly handled therein