+ installx: only overwrite if changed, with flag to force
for a long time, it's been troublesome that we unconditionally overwrote
all files on install, even if they haven't changed (ie we only changed
or added one script).  this caused the installed files to lose
information about when they were last updated.  now, for installx (but
not yet installrc), we'll check that either the symlink is the same (for
link) or the file has the same contents (for script).

this will also keep track and display them as skipped in the stat print
at end.

user can force the install with --unchanged flag

installrc is more complicated, because it's installing symlinks, and it
supports multi-level links (unlinke installx), which furthermore will be
linking somewhere else (ie to the source directory) than the actually
installed symlink, so the targeting cannot use a direct
readlink-chain-compare.  that will take a bit more thinking and be
implemented later