/ field: main: flush stdout/stderr to avoid broken pipe msgs
it's amazing that python still needs this, the bug was opened in 2011
https://github.com/python/cpython/issues/55589
nee https://bugs.python.org/issue11380

implemented here first, but we copied it to devskel also just prior, as
this should be standard.  sad that it's still required.  only tested on
python 3.9 though but someone would have touched the bug probably if it
were fixed