/ field: select: rather than timeout 0, check if stdin isatty(3)
when checking for input, we tested whether stdin was ready to read
immediately.  sometimes this is not the case such as when the program
writing to us in a pipeline does some processing first, we will get to
the select(2) before any input is ready.  then since we had set timeout
zero, the call returns immediately and we are forced to conclude a usage
error, but this is incorrect.  this leads to naive case like "echo foo |
field" working correctly but "(sleep 1; echo foo) | field" fails.  quite
a few programs we would run in a pipeline do not have output available
to write yet by the time we are doing select().  isatty() is how
typicaly unix filter programs usually do it...