/ field: dmain: finally: move broken pipe flush to dmain for outer except
if this isn't outside the root driver function, the last exception is
eaten and we never get a printout.  we have to catch it ourselves.  so
this block belongs in dmain rather than main

we have to re-raise SystemExit also which previously would not have been
caught when we use sys.exit() but now we will get it, but don't want to
print exception in that case because it's actually normal
operation/expected and was requested by the user