- field: process_args: do -r/--regex check in one place only
the --regex option is only useful in the case that positional args were
given, because it actually only modifies the sole case of
positionally-specified ifs (to make it a regex).  so we can move move
the check for --regex to the same place that we determine if positional
args were used, and after positional arg processing, we no longer care
about --regex because there's already a barrier check for both options
and positional args being specified.

this means we only need to check if field options were specified in one
place, so there's no longer a need for 'havefopts' either