/ field: field: brokenpipe: exit rather than pass to avoid hang
we caught broken pipe in 8c5948d3 and 'pass'ed it, considering this a
normal event such as when piping to 'head'.  while this is true, the
correct behavior is to stop the program at that point because there is
no reader left to consume our output and this actually signals an exit
condition.  however, it's not an error per se, so we return success.