+ field: noendrec: new option not to output the final trailing ORS
we usually always output a trailing ors, even though technically this
would create an empty field as the last record.  in reality ORS (which,
afaik, came from awk) is a misnomer: they really are more correctly
designated "record terminators" rather than "record separators."

however, in some use cases -- example, taking a list of PIDs one per line
and changing their record separator to comma, then giving the resulting
list to 'ps', which balks on trailing commas -- we will want to set the
ors to ',', but we do *not* want a trailing comma on the last record.

so we create an option '-N/--noendrec' that will omit this trailing one,
for this special case