+ field: output: always give a trailing newline when stdout isatty
when the output is a terminal, we'll always want a newline sent, as this
is the expectation shells have of programs which output, when viewing
results.  if output is a pipeline or file, we do nothing, but if a tty,
we add a newline.

if the ors is already a newline -- the usual case -- we just let that
form the final program output separator, and do nothing.  when ors is
not a newline, we fabricate one and add it to the output.