+ hostfill: filter output through sort -V, groups stuff better
we want sorted output so things are grouped better, and we want version
sort so network numbers look correct

looked into doing the sort within the python and, while it was easy
enough to do version sort with a list of numbers in a specific format,
there is no general way to do this with strings in arbitrary formats
that might happen to contain a version, without implementing the sort
myself.  no reason not to just use the system sort utility, this is not
a time-sensitive script and plenty of time is already taken by the
hostname lookups anyways.  there is no system without 'sort' that isn't
some application container and we wouldn't be running this script in
there anyways so safe to assume it