+ hostfill: add argument processing and make sort an option
we might not always want this.  for example it will not look good when
"ip addr ls | hostfill" if it's sorted, however "ip route ls | hostfill"
does look better with the sort.  not sure what the default should be,
we'll leave it off for now and can always come back and fix this later

this is now only saving 7 keystrokes: " -s" " | sort -V" (well, that's
enough to justify implementing it anyways, since we have been using this
a lot)