+ lensort: new script that sorts its input by length, with some options
the use case was coming up with short identifiers by using something
like "thes foo" ie our thesaurus alias, which prints columnated text
synonyms or related words.  we wanted it in one column and sorted by
length, with a cutoff value.  and it needed to handle the sometimes
phrases with space in them, hence the -2 option

we were using an impromptu command line to do this but it's better to
have it in a script and put it in our utility collection