+ yamldump: string-passthrough ansible yaml vault-encrypted fields
it would crash on this !vault construct before.  we don't actually
decrypt them, we just pass through as strings.  but at least it "works"
and will dump the file as a python structure, which is what we want
(usually we're testing syntax or confirming our understanding of the
parse tree).